@@ -1,4 +1,4 @@
-var requireIndex = require("requireindex");
+const requireIndex = require('requireindex');
 
 // import all rules in lib/rules
-module.exports.rules = requireIndex(__dirname + "/rules");
+module.exports.rules = requireIndex(`${__dirname}/rules`);
